+++ openpecha/formatters/layers.py
@@ -2,8 +2,6 @@
 This module contains format variable for all the annotations
 """
 
-from collections import namedtuple
-from enum import Enum
 
 __all__ = [
     "Layer",
@@ -30,30 +28,6 @@ __all__ = [
     "Footnote",
 ]
 
-
-class AnnType(Enum):
-    book_title = "BookTitle"
-    sub_title = "SubTitle"
-    book_number = "BookNumber"
-    poti_title = "PotiTitle"
-    author = "Author"
-    chapter = "Chapter"
-
-    topic = "Text"
-    sub_topic = "SubText"
-
-    pagination = "Pagination"
-    citation = "Citation"
-    correction = "Correction"
-    error_candidate = "ErrorCandidate"
-    peydurma = "Peydurma"
-    pedurma_note = "PedurmaNote"
-    sabche = "Sabche"
-    tsawa = "Tsawa"
-    yigchung = "Yigchung"
-    archaic = "Archaic"
-    durchen = "Durchen"
-    footnote = "Footnote"
 
 class _attr_names:
     # Layer